Move `test` directory to `form_manager` directory
The frontend directory lives in the same repo at the root level.
So it is better to keep things isolated from each other for clarity's
sake.
Tests as part of application code is a layout that is still officially
supported:
(https://docs.pytest.org/en/stable/explanation/goodpractices.html#tests-as-part-of-application-code)